LayoutTests/imported/w3c/web-platform-tests/cssom-view/scrollintoview-expected.txt


PASS scrollIntoView() starting at left,top 
PASS scrollIntoView() starting at left,bottom 
PASS scrollIntoView() starting at right,top 
PASS scrollIntoView() starting at right,bottom 
PASS scrollIntoView(true) starting at left,top 
PASS scrollIntoView(true) starting at left,bottom 
PASS scrollIntoView(true) starting at right,top 
PASS scrollIntoView(true) starting at right,bottom 
PASS scrollIntoView(false) starting at left,top 
PASS scrollIntoView(false) starting at left,bottom 
PASS scrollIntoView(false) starting at right,top 
PASS scrollIntoView(false) starting at right,bottom 
PASS scrollIntoView(undefined) starting at left,top 
PASS scrollIntoView(undefined) starting at left,bottom 
PASS scrollIntoView(undefined) starting at right,top 
PASS scrollIntoView(undefined) starting at right,bottom 
FAIL scrollIntoView(null) starting at left,top assert_approx_equals: scrollY expected 4000 +/- 0.5 but got 3600
FAIL scrollIntoView(null) starting at left,bottom assert_approx_equals: scrollY expected 4000 +/- 0.5 but got 3600
FAIL scrollIntoView(null) starting at right,top assert_approx_equals: scrollY expected 4000 +/- 0.5 but got 3600
FAIL scrollIntoView(null) starting at right,bottom assert_approx_equals: scrollY expected 4000 +/- 0.5 but got 3600
PASS scrollIntoView({}) starting at left,top 
PASS scrollIntoView({}) starting at left,bottom 
PASS scrollIntoView({}) starting at right,top 
PASS scrollIntoView({}) starting at right,bottom 
FAIL scrollIntoView({block: "center", inline: "center"}) starting at left,top assert_approx_equals: scrollX expected 3700 +/- 0.5 but got 3400
FAIL scrollIntoView({block: "center", inline: "center"}) starting at left,bottom assert_approx_equals: scrollX expected 3700 +/- 0.5 but got 3400
FAIL scrollIntoView({block: "center", inline: "center"}) starting at right,top assert_approx_equals: scrollX expected 3700 +/- 0.5 but got 4000
FAIL scrollIntoView({block: "center", inline: "center"}) starting at right,bottom assert_approx_equals: scrollX expected 3700 +/- 0.5 but got 4000
FAIL scrollIntoView({block: "start", inline: "start"}) starting at left,top assert_approx_equals: scrollX expected 4000 +/- 0.5 but got 3400
FAIL scrollIntoView({block: "start", inline: "start"}) starting at left,bottom assert_approx_equals: scrollX expected 4000 +/- 0.5 but got 3400
PASS scrollIntoView({block: "start", inline: "start"}) starting at right,top 
PASS scrollIntoView({block: "start", inline: "start"}) starting at right,bottom 
FAIL scrollIntoView({block: "end", inline: "end"}) starting at left,top assert_approx_equals: scrollY expected 3600 +/- 0.5 but got 4000
FAIL scrollIntoView({block: "end", inline: "end"}) starting at left,bottom assert_approx_equals: scrollY expected 3600 +/- 0.5 but got 4000
FAIL scrollIntoView({block: "end", inline: "end"}) starting at right,top assert_approx_equals: scrollX expected 3400 +/- 0.5 but got 4000
FAIL scrollIntoView({block: "end", inline: "end"}) starting at right,bottom assert_approx_equals: scrollX expected 3400 +/- 0.5 but got 4000
FAIL scrollIntoView({block: "nearest", inline: "nearest"}) starting at left,top assert_approx_equals: scrollY expected 3600 +/- 0.5 but got 4000
PASS scrollIntoView({block: "nearest", inline: "nearest"}) starting at left,bottom 
FAIL scrollIntoView({block: "nearest", inline: "nearest"}) starting at right,top assert_approx_equals: scrollY expected 3600 +/- 0.5 but got 4000
PASS scrollIntoView({block: "nearest", inline: "nearest"}) starting at right,bottom 
